// src/store/main/system/types.ts
/*
 * @Author: Crayon 3037686283@qq.com
 * @Date: 2023-03-03 10:36:56
 * @LastEditors: Crayon 3037686283@qq.com
 * @LastEditTime: 2023-03-08 18:52:29
 * @FilePath: \manager_vue3\manager_-system\src\store\main\system\types.ts
 * @Description: 这是默认设置,请设置`customMade`, 打开koroFileHeader查看配置 进行设置: https://github.com/OBKoro1/koro1FileHeader/wiki/%E9%85%8D%E7%BD%AE
 */
export interface ISystemState {
  userList: any[]
  userCount: number
  roleList: any[]
  roleCount: number
}